Separate v1 tests from current state of play
