+++ e2e/pages/manage/v1/createPlacementPage.ts
@@ -1,4 +1,4 @@
-import { BasePage } from '../basePage'
+import { BasePage } from '../../basePage'
 
 export class CreatePlacementPage extends BasePage {
   async chooseAp() {
