+++ e2e/tests/apply/withdrawals.spec.ts
@@ -1,15 +1,15 @@
-import { test } from '../test'
+import { test } from '../../test'
 import {
   enterAndConfirmCrn,
   startAnApplication,
   visitDashboard,
   withdrawAnApplicationBeforeSubmission,
-} from '../steps/apply'
+} from '../../steps/apply'
 
-import { ListPage } from '../pages/apply'
+import { ListPage } from '../../pages/apply'
 
-import { ShowPage } from '../pages/apply/showPage'
-import { signIn } from '../steps/signIn'
+import { ShowPage } from '../../pages/apply/showPage'
+import { signIn } from '../../steps/signIn'
 
 test('Withdraw an application before submission', async ({ page, person, userWithoutRoles }) => {
   await signIn(page, userWithoutRoles)
