+++ e2e/tests/manage/out-of-service-beds.spec.ts
